+++ src/ui/overlay/data_marker.hpp
@@ -23,17 +23,19 @@ class Axes3D;
 // A persistent data marker pinned to a specific data point.
 struct DataMarker
 {
-    float         data_x      = 0.0f;
+    // Absolute data x (includes series x_offset); double so epoch-scale
+    // timestamps keep sub-second precision.
+    double        data_x      = 0.0;
     float         data_y      = 0.0f;
     float         data_z      = 0.0f;
     const Series* series      = nullptr;
     size_t        point_index = 0;
     Color         color       = colors::white;
-    std::string   series_label;              // series name shown in datatip
-    const Axes*   axes        = nullptr;     // owning 2D axes (for multi-subplot)
-    const Axes3D* axes3d     = nullptr;     // owning 3D axes
-    float         dy_dx       = 0.0f;        // finite-difference derivative
-    bool          dy_dx_valid = false;       // true when derivative was computed
+    std::string   series_label;            // series name shown in datatip
+    const Axes*   axes        = nullptr;   // owning 2D axes (for multi-subplot)
+    const Axes3D* axes3d      = nullptr;   // owning 3D axes
+    float         dy_dx       = 0.0f;      // finite-difference derivative
+    bool          dy_dx_valid = false;     // true when derivative was computed
     bool          is_3d       = false;
 };
 
@@ -44,7 +46,7 @@ class DataMarkerManager
    public:
     DataMarkerManager() = default;
 
-    void add(float         data_x,
+    void add(double        data_x,
              float         data_y,
              const Series* series,
              size_t        index,
@@ -59,7 +61,7 @@ class DataMarkerManager
 
     // Toggle a data label: if a marker already exists at this (series, index),
     // remove it; otherwise add a new one. Returns true if a label was added.
-    bool toggle_or_add(float         data_x,
+    bool toggle_or_add(double        data_x,
                        float         data_y,
                        const Series* series,
                        size_t        index,
@@ -74,11 +76,11 @@ class DataMarkerManager
                 size_t        index,
                 const Axes3D* axes3d);
     bool toggle_or_add_3d(float         data_x,
-                            float         data_y,
-                            float         data_z,
-                            const Series* series,
-                            size_t        index,
-                            const Axes3D* axes3d);
+                          float         data_y,
+                          float         data_z,
+                          const Series* series,
+                          size_t        index,
+                          const Axes3D* axes3d);
 
     const std::vector<DataMarker>& markers() const { return markers_; }
     size_t                         count() const { return markers_.size(); }
@@ -87,8 +89,8 @@ class DataMarkerManager
     // When filter_axes is non-null, only markers belonging to that axes are drawn.
     // When dl is non-null, draws into the given draw list instead of GetForegroundDrawList().
     void draw(const Rect& viewport,
-              float       xlim_min,
-              float       xlim_max,
+              double      xlim_min,
+              double      xlim_max,
               float       ylim_min,
               float       ylim_max,
               float       opacity     = 1.0f,
@@ -102,8 +104,8 @@ class DataMarkerManager
     int hit_test(float       screen_x,
                  float       screen_y,
                  const Rect& viewport,
-                 float       xlim_min,
-                 float       xlim_max,
+                 double      xlim_min,
+                 double      xlim_max,
                  float       ylim_min,
                  float       ylim_max,
                  float       radius_px   = 10.0f,
@@ -111,10 +113,10 @@ class DataMarkerManager
 
     // Draw / hit-test compact pinned labels on 3D axes (camera-projected).
     void draw_3d(const Axes3D& axes3d, float opacity = 1.0f, ImDrawList* dl = nullptr);
-    int  hit_test_3d(float             screen_x,
-                     float             screen_y,
-                     const Axes3D&     axes3d,
-                     float             radius_px = 10.0f) const;
+    int  hit_test_3d(float         screen_x,
+                     float         screen_y,
+                     const Axes3D& axes3d,
+                     float         radius_px = 10.0f) const;
 
    private:
     std::vector<DataMarker> markers_;
@@ -125,16 +127,15 @@ class DataMarkerManager
     int find_duplicate(const Series* series, size_t point_index) const;
 
     // Convert data coordinates to screen coordinates
-    static void data_to_screen(float       data_x,
+    static void data_to_screen(double      data_x,
                                float       data_y,
                                const Rect& viewport,
-                               float       xlim_min,
-                               float       xlim_max,
+                               double      xlim_min,
+                               double      xlim_max,
                                float       ylim_min,
                                float       ylim_max,
                                float&      screen_x,
                                float&      screen_y);
-
 };
 
 }   // namespace spectra
